Changed various things in Level class - modified Tile.obj.model (reducing its size by 2 px each XY) - the size of the level is now dynamic - removed linearized array as it isn't necessary - added second level by expanding level var to an array of Feld - added OutOfBound() to prevent null error exceptions

+++ src/Engine/Examples/MyTestGame/Feld.cs
@@ -10,8 +10,8 @@ namespace Examples.MyTestGame
         public FieldTypes Type;
         public FieldStates State;
 
-        public byte X;
-        public byte Y;
+        public int X;
+        public int Y;
         private static Mesh _feldMesh;
 
         public enum FieldTypes
